drapixai_ai/requirements.txt
torch==2.4.0
torchvision==0.19.0
xformers==0.0.27.post2

fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.8
rq==1.16.2
python-multipart==0.0.9
requests==2.32.3
boto3==1.34.131

accelerate==0.34.2
transformers==4.46.3
diffusers==0.31.0
huggingface_hub>=0.34.0,<2.0
safetensors
peft==0.17.0
einops==0.7.0

pillow==10.3.0
numpy==1.26.4
opencv-python==4.10.0.84
scipy==1.13.1
scikit-image==0.24.0
matplotlib==3.9.1
PyYAML==6.0.1
tqdm==4.66.4
rembg==2.0.57
onnxruntime==1.23.2

fvcore==0.1.5.post20221221
cloudpickle==3.0.0
omegaconf==2.3.0
pycocotools==2.0.8
av==12.3.0
